refactor ValidatorModuleTest
