--- zheng-validator-bval/src/test/java/com/github/zhengframework/validator/bval/DummyException.java
@@ -1,15 +1,0 @@
-package com.github.zhengframework.validator.bval;
-
-public class DummyException extends Exception {
-
-  private static final long serialVersionUID = 1L;
-
-  public DummyException(Throwable cause) {
-    super(cause);
-  }
-
-  public DummyException(String message, Throwable cause) {
-    super(message, cause);
-  }
-
-}
