--- zheng-validator-bval/src/test/java/com/github/zhengframework/validator/bval/Insert.java
@@ -1,7 +1,0 @@
-
-package com.github.zhengframework.validator.bval;
-
-
-public interface Insert {
-
-}
