--- zheng-validator-bval/src/test/java/com/github/zhengframework/validator/bval/Update.java
@@ -1,6 +1,0 @@
-package com.github.zhengframework.validator.bval;
-
-
-public interface Update {
-
-}
