--- zheng-validator-hibernate/src/test/java/com/github/zhengframework/validator/hibernate/DummyException.java
@@ -1,11 +1,0 @@
-package com.github.zhengframework.validator.hibernate;
-
-public class DummyException extends Exception {
-
-  private static final long serialVersionUID = 1L;
-
-  public DummyException(String message, Throwable cause) {
-    super(message, cause);
-  }
-
-}
