--- zheng-validator-hibernate/src/test/java/com/github/zhengframework/validator/hibernate/Insert.java
@@ -1,7 +1,0 @@
-
-package com.github.zhengframework.validator.hibernate;
-
-
-public interface Insert {
-
-}
