--- zheng-validator-hibernate/src/test/java/com/github/zhengframework/validator/hibernate/Update.java
@@ -1,6 +1,0 @@
-package com.github.zhengframework.validator.hibernate;
-
-
-public interface Update {
-
-}
