+++ zheng-validator-test/pom.xml
@@ -2,30 +2,16 @@
 <project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
   xmlns="http://maven.apache.org/POM/4.0.0"
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <artifactId>zheng-validator-bval</artifactId>
+  <artifactId>zheng-validator-test</artifactId>
   <dependencies>
+    <dependency>
+      <artifactId>junit</artifactId>
+      <groupId>junit</groupId>
+      <scope>compile</scope>
+    </dependency>
     <dependency>
       <artifactId>zheng-validator</artifactId>
       <groupId>com.github.zhengframework</groupId>
-    </dependency>
-    <dependency>
-      <artifactId>bval-jsr</artifactId>
-      <groupId>org.apache.bval</groupId>
-      <version>2.0.3</version>
-    </dependency>
-    <dependency>
-      <artifactId>zheng-test</artifactId>
-      <groupId>com.github.zhengframework</groupId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <artifactId>slf4j-simple</artifactId>
-      <groupId>org.slf4j</groupId>
-    </dependency>
-    <dependency>
-      <artifactId>zheng-validator-test</artifactId>
-      <groupId>com.github.zhengframework</groupId>
-      <scope>test</scope>
     </dependency>
   </dependencies>
   <dependencyManagement>
