+++ zheng-validator-test/src/main/java/com/github/zhengframework/validator/Insert.java
@@ -1,0 +1,7 @@
+
+package com.github.zhengframework.validator;
+
+
+public interface Insert {
+
+}
