+++ zheng-validator-test/src/main/java/com/github/zhengframework/validator/Update.java
@@ -1,0 +1,6 @@
+package com.github.zhengframework.validator;
+
+
+public interface Update {
+
+}
